+++ makeprior/makeprior.cpp
@@ -110,7 +110,7 @@ void merge_sorted_runs( indri::file::File& out, std::vector<std::string>& inputs
   std::priority_queue< MergeFile > files;
 
   // open files
-  for( int i=0; i<inputs.size(); i++ ) {
+  for( size_t i=0; i<inputs.size(); i++ ) {
     indri::file::File* file = new indri::file::File;
     file->openRead( inputs[i] );
     indri::file::SequentialReadBuffer* buffer = new indri::file::SequentialReadBuffer( *file, 512*1024 );
@@ -222,7 +222,7 @@ void sort_file( indri::file::File& out, indri::file::File& in, size_t memory, in
   in.close();
   merge_sorted_runs( out, temporaries, totalDocuments );
   
-  for( int i=0; i<temporaries.size(); i++ ) {
+  for( size_t i=0; i<temporaries.size(); i++ ) {
     lemur_compat::remove( temporaries[i].c_str() );
   }
 }
@@ -302,8 +302,8 @@ void compress_file( indri::file::File& out, indri::file::File& in, const std::ma
   indri::file::SequentialReadBuffer* inb = new indri::file::SequentialReadBuffer( in, 512*1024 );
   indri::file::SequentialWriteBuffer* outb = new indri::file::SequentialWriteBuffer( out, 512*1024 );
   
-  int itemCount = 0;
-  int tableCount = values.size();
+  UINT32 itemCount = 0;
+  UINT32 tableCount = values.size();
   
   inb->read( &itemCount, sizeof(UINT32) );
   inb->read( &tableCount, sizeof(UINT32) );
@@ -316,13 +316,13 @@ void compress_file( indri::file::File& out, indri::file::File& in, const std::ma
   outb->write( &tableCount, sizeof(UINT32) );
   
   // write table
-  for( int i=0; i<tableCount; i++ ) {
+  for( UINT32 i=0; i<tableCount; i++ ) {
     double value = (double) inverted[i];
     outb->write( &value, sizeof(double) );
   }
   
   // write indexes  
-  for( int i=0; i<itemCount; i++ ) {
+  for( UINT32 i=0; i<itemCount; i++ ) {
     double value;
     inb->read( &value, sizeof(double) );
   
@@ -405,7 +405,7 @@ int main( int argc, char** argv ) {
     indri::collection::Repository::index_state indexes = _repository->indexes();
     INT64 documentCount = 0;
   
-    for( int i=0; i<indexes->size(); i++ ) {
+    for( size_t i=0; i<indexes->size(); i++ ) {
       indri::thread::ScopedLock lock( (*indexes)[i]->statisticsLock() );
       documentCount += (*indexes)[i]->documentCount();
     }
