+++ src/Combiner.cpp
@@ -71,7 +71,7 @@ void indri::parse::Combiner::_openWriteBuckets( std::vector<std::stringstream*>&
 }
 
 void indri::parse::Combiner::_closeWriteBuckets( std::vector<std::stringstream*>& buffers, std::vector<std::ofstream*>& buckets ) {
-  for( unsigned int i=0; i<buckets.size(); i++ ) {
+  for( size_t i=0; i<buckets.size(); i++ ) {
     _flushWriteBuffer( buffers, buckets, true, i );
     delete buffers[i];
     buckets[i]->close();
@@ -80,7 +80,7 @@ void indri::parse::Combiner::_closeWriteBuckets( std::vector<std::stringstream*>
 }
 
 void indri::parse::Combiner::_flushWriteBuffers( std::vector<std::stringstream*>& buffers, std::vector<std::ofstream*>& buckets, bool force ) {
-  for( unsigned int i=0; i<buckets.size(); i++ ) {
+  for( size_t i=0; i<buckets.size(); i++ ) {
     _flushWriteBuffer( buffers, buckets, force, i );
   }
 }
@@ -331,7 +331,7 @@ std::stringstream* combiner_flush_stream( std::stringstream* buffer, std::ofstre
 }
 
 void indri::parse::Combiner::closeBuckets() {
-  for( unsigned int i=0; i<_docBuckets.size(); i++ ) {
+  for( size_t i=0; i<_docBuckets.size(); i++ ) {
     _docBucketFiles[i]->close();
     delete _docBuckets[i];
     delete _docBucketFiles[i];
@@ -339,7 +339,7 @@ void indri::parse::Combiner::closeBuckets() {
   _docBuckets.clear();
   _docBucketFiles.clear();
 
-  for( unsigned int i=0; i<_linkBuckets.size(); i++ ) {
+  for( size_t i=0; i<_linkBuckets.size(); i++ ) {
     _linkBucketFiles[i]->close();
     delete _linkBuckets[i];
     delete _linkBucketFiles[i];
@@ -403,7 +403,7 @@ void indri::parse::Combiner::_writeCorpusTable( UrlEntryVectorTable& corpusTable
 
     // write to it
     // TODO: note that this may cause disk fragmentation
-    for( unsigned int i=0; i<entryVec->size(); i++ ) {
+    for( size_t i=0; i<entryVec->size(); i++ ) {
       url_entry* entry = (*entryVec)[i];
       std::string number = i64_to_string( entry->linkCount );
