+++ src/DocumentVector.cpp
@@ -32,7 +32,7 @@ void indri::api::DocumentVector::_init( indri::index::Index* index, const indri:
   std::map<int, int> termIDMap;
   _stems.push_back( "[OOV]" );
 
-  for( unsigned int i=0; i<terms.size(); i++ ) {
+  for( size_t i=0; i<terms.size(); i++ ) {
     if( terms[i] == 0 ) {
       _positions.push_back(0);
     } else {
@@ -65,7 +65,7 @@ void indri::api::DocumentVector::_init( indri::index::Index* index, const indri:
     }
   }
 
-  for( unsigned int i=0; i<fields.size(); i++ ) {
+  for( size_t i=0; i<fields.size(); i++ ) {
     Field f;
 
     f.name = index->field(fields[i].id);
